FIX: #1 Add 'dotnet add package' command to include FoodService.Nugget.Models
This commit updates the YAML workflow file to include the 'dotnet add package FoodService.Nugget.Models' command right after the 'dotnet restore' command. This ensures that the 'FoodService.Nugget.Models' package is installed in the project before compilation, testing, and publishing, thus resolving the error of not finding the package during restoration.