Updated project to use SqlServer data context
This commit includes a major refactoring of the project to use a SqlServer specific data context. The project reference in `FoodServiceAPI.Core.csproj` was changed to `FoodServiceAPI.Data.SqlServer.csproj`. The `IGenericRepository` interface and its implementations were removed and replaced with SqlServer specific implementations. The `FoodServiceAPI.Data.csproj` file was removed, and the `FoodServiceAPI.Data.SqlServer.csproj` file was added. Several files were updated to use the `FoodServiceAPI.Data.SqlServer` namespaces. The `CpfCnpjValidator.cs` file was moved to the `FoodServiceAPI.Shared.Validator` namespace. The `FoodServiceAPI.sln` file was updated to reference the `FoodServiceAPI.Data.SqlServer` project. The `AppDbContext.cs` file was moved to the `FoodServiceAPI.Data.SqlServer.Context` namespace. New repository files were added in the `FoodServiceAPI.Data.SqlServer.Repository` namespace. The `SqlServerConfig.cs` file was added with methods for configuring the database and updating the database migration. The `GenericRepository.cs` and `IGenericRepository.cs` files were added, providing a generic implementation of CRUD operations.
Updated FoodServiceAPI to use SQL Server and revised data types

This commit includes a significant overhaul of the FoodServiceAPI. The primary change is the switch from MySQL to SQL Server, which involved updating the `FoodService.Models` package, replacing the `Pomelo.EntityFrameworkCore.MySql` package with `Microsoft.EntityFrameworkCore.SqlServer`, and modifying the connection string in `appsettings.json` and `DatabaseConfig.cs`.

The `SiteSettingsConfiguration` class was removed and replaced with `SiteSettingsConfig` for configuring the `SiteSettings` entity. The initial database migration files `20240605162306_Initial.Designer.cs` and `20240605162306_Initial.cs` were removed, and a new migration for the initial creation of the database schema was created.

Data types were revised to match SQL Server data types, including changing price values from decimal to double in `AppDbContext.cs` and various changes in `AppDbContextModelSnapshot.cs`. The maximum identifier length was increased from 64 to 128.

Several methods were replaced to accommodate the switch to SQL Server, and the `ConfigureDatabase` method in `DatabaseConfig.cs` and `Program.cs` was updated to use SQL Server.

The `SiteSettingsConfig.cs` file was added, and the `AppDbContextModelSnapshot.cs` file was updated to reflect the changes in the database schema. The expiration dates and prices of some products were updated in `AppDbContextModelSnapshot.cs`.

Finally, a new migration for the initial creation of the database schema was created, with several new tables added and initial data inserted into the `Ingredient` and `Product` tables, and a single row of settings into the `SiteSettings` table.