DEV🛠️: Update GitHub Actions workflow and solution file references
Updated the GitHub Actions workflow in `azure_foodservice-api.yml` to include a new step for adding GitHub Packages NuGet source and restoring project dependencies. Also, updated the deployment step to include a publish profile secret for Azure Web App deployment. In the `FoodServiceAPI.sln` file, updated the reference to reflect the renaming of the workflow file in the 'Solution Items' project section.